resnet_webapp/requirements.txt
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.0.9
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
comm==0.1.4
cryptography==41.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.8
django-cors-headers==4.3.1
django-pwa==1.1.0
djangorestframework==3.14.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fqdn==1.5.1
# gmpy2==2.1.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.12.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.6.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
# mkl-fft==1.3.1
# mkl-random==1.2.1
# mkl-service==2.4.0
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.2
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.42
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
# pyobjc-core==10.1
# pyobjc-framework-Cocoa==10.1
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
Send2Trash==1.8.2
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tornado==6.3.3
traitlets==5.14.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
typing-utils==0.1.0
tzdata==2023.4
uri-template==1.3.0
urllib3==1.26.18
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wheel==0.41.2
zipp==3.17.0
